Fix initial color picker color when loading page
The colors configured for the different modes are persisted on the lamp.
When re-opening the user interface, the color picker did not show the
currently configured color.

The JS fronted already has the currently set color available, the color
picker simply has not been updated to the correct color.

This commit fixes the "Color Wipe" picker only to make it easier to
review it. Subsequent commits will fix the other pickers.